requirements.txt
fastf1==3.4.0
pandas>=2.0.0
numpy>=1.24.0
kafka-python>=2.0.2
azure-cosmos==4.5.1
azure-functions==1.17.0
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0
python-dotenv==1.0.0
pydantic==2.5.2
asyncio-mqtt==0.16.1
redis==5.0.1 